poi radio unit testing
